gripper state normalized

+++ xarm_moveit_servo/include/xarm_moveit_servo/xarm_joystick_input.h
@@ -16,6 +16,7 @@
 #include <moveit_msgs/msg/planning_scene.hpp>
 #include "xarm_moveit_servo/gripper_controller.h"
 #include "std_msgs/msg/int32.hpp"
+#include "std_msgs/msg/float64.hpp"
 
 
 namespace xarm_moveit_servo
@@ -42,7 +43,7 @@ private:
     rclcpp::Publisher<geometry_msgs::msg::TwistStamped>::SharedPtr twist_pub_;
     rclcpp::Publisher<control_msgs::msg::JointJog>::SharedPtr joint_pub_;
     rclcpp::Publisher<std_msgs::msg::Int32>::SharedPtr gripper_command_pub_;
-    rclcpp::Publisher<std_msgs::msg::Int32>::SharedPtr gripper_state_pub_;
+    rclcpp::Publisher<std_msgs::msg::Float64>::SharedPtr gripper_state_pub_;
     rclcpp::Publisher<moveit_msgs::msg::PlanningScene>::SharedPtr collision_pub_;
     rclcpp::Client<std_srvs::srv::Trigger>::SharedPtr servo_start_client_;
